+++ src/kadrx/PulseData.h
@@ -7,6 +7,7 @@
 /// KaMerge object.
 
 #include <sys/time.h>
+#include <sys/types.h>
 
 class PulseData {
 
@@ -24,21 +25,21 @@ public:
 
   // set the data
   
-  void set(long long pulseSeqNum,
+  void set(int64_t pulseSeqNum,
            time_t timeSecs,
            int nanoSecs,
            int channel,
            int gates,
-           const unsigned short *iq);
+           const int16_t *iq);
 
   // get methods
 
-  long long getPulseSeqNum() const { return _pulseSeqNum; }
+  int64_t getPulseSeqNum() const { return _pulseSeqNum; }
   time_t getTimeSecs() const { return _timeSecs; }
   int getNanoSecs() const { return _nanoSecs; }
   int getChannel() const { return _channel; }
   int getGates() const { return _gates; }
-  const unsigned short *getIq() const { return _iq; }
+  const int16_t *getIq() const { return _iq; }
     
 private:
 
@@ -46,7 +47,7 @@ private:
    * pulse sequence number since start of ops
    */
   
-  long long _pulseSeqNum;
+  int64_t _pulseSeqNum;
   
   /**
    * time of pulse - seconds and nano-secs
@@ -72,7 +73,7 @@ private:
    * IQ data
    */
 
-  unsigned short *_iq;
+  int16_t *_iq;
 
   // functions
